Sherlock.jl

Sherlock.Detective

Sherlock.MIMO

Sherlock.abstracttypes

Sherlock.functions

Sherlock.functiontype_edges

Sherlock.inquire

Sherlock.ismacro

Sherlock.magnify

Sherlock.nothing_is_val

Sherlock.sherlock_UI

Sherlock.sherlockplot

Sherlock.trim_type

Sherlock.types

Sherlock.typetree

Sherlock.typetype_edges

Sherlock.undefined

Sherlock.Detective — Method

Detective( moduleinst::Module )

Instantiate Detective object via a module.

Sherlock.MIMO — Method

MIMO( d::Detective, s::Symbol)

Returns a subgraph plot of the first order interactions of a given object in a Detective instance.

Sherlock.abstracttypes — Method

abstracttypes(d::Detective)::Vector{Symbol}

Returns a list of abstract types of a module in a Detective object.

Sherlock.functions — Method

functions(d::Detective)::Vector{Symbol}

Returns a list of functions of a module in a Detective object.

Sherlock.functiontype_edges — Method

functiontype_edges( d::Detective )

Discovers which types are called from the available functions in a package. It adds that information to a Detective instance's internal graph object as an inplace operation.

Sherlock.inquire — Method

inquire( d::Detective, s::Union{Symbol,String} )

This function will search a loaded module for an instance or method of symbol s. If it finds it, it will, return an enumeration of its category. This is useful for quickly categorizing an item in a struct for display purposes.

Sherlock.ismacro — Method

ismacro(s::Symbol)

checks the first character of an input symbol to see if it contains a @ or not.

Sherlock.magnify — Method

magnify( d::Detective, s::Symbol)

Returns a subgraph plot of the first order interactions of a given object in a Detective instance.

Sherlock.nothing_is_val — Function

nothing_is_val( x::Union{Nothing,Int}, val = Inf )

if x is a type nothing return the alternative, val param. Otherwise returns x. Note: this is specifically used for findfirst() failures internally.

Sherlock.sherlock_UI — Method

sherlock_UI()

This is a convenience function to create a Blink window which hosts an Interact UI. The intention of the UI is to explore a module (typed in by a user), and display some of the knowledge obtained about it graphically.

Sherlock.sherlockplot — Method

sherlockplot(d::Detective)

Returns a LightGraphs plot recipe of a Detective instances internal graph.

Sherlock.trim_type — Method

trim_type( s::Symbol )::Symbol

Trims a type Symbol to the first {, or . Also takes the last subtype after a ..

Sherlock.types — Method

types(d::Detective)::Vector{Symbol}

Returns a list of types of a module in a Detective object.

Sherlock.typetree — Method

typetree( s::Symbol )

Returns a subgraph type tree plot of a given type

This is literally a wrapper for basic GraphRecipes functionality.

Sherlock.typetype_edges — Method

typetype_edges( d::Detective )

Discovers first order relationships between types in a package. It adds that information to a Detective instance's internal graph object as an inplace operation.

Sherlock.undefined — Method

undefined(d::Detective)::Vector{Symbol}

Returns a list of undefined objects in a module in a Detective object. This could be exported items that don't exist, or objects without explicit types, until compile time.