ClusterScripts.jl API

ClusterScripts.ResultsLazyLoader — Type

Struct to view into results files, loading full results only if directly accessed.

If parameters or derivedquantities are modified, remember to call save!() to update the file. Modifications to results outside those made by `concatenateresults!` will not be saved.

ClusterScripts.SimulationFile — Type

Struct to hold file paths and provide some basic functionality for working with them.

Base.setindex! — Method

Base.setindex!(loader::ResultsLazyLoader, val, i::Int)

Warning: This overwrites the entire results group! use loader[i]=append!(deepcopy(loader[i]), X) or similar to append to existing data.

ClusterScripts.build_job_queue — Method

build_job_queue(fixed_parameters::Dict, variables::Dict, postprocessing_function::Function)

Returns a Vector of all unique combinations of values in variables merged with fixed_parameters. By specifying a postprocessing_function, further actions can be performed each of the elements in the resulting vector.

ClusterScripts.build_job_queue — Method

build_job_queue(fixed_parameters::Dict, variables::Dict)

Returns a Vector of all unique combinations of values in variables merged with fixed_parameters. Each key in variables should be a list of possible values for that parameter. (Trivially, a length 1 list)

ClusterScripts.convert_to_grouped_jld2 — Method

convert_to_grouped_jld2(filename, results_data, simulation_queue; trajectories_key = "trajectories")

Converts the results format of an ungrouped JLD2 file to the grouped format and adds the simulation parameters from a simulation queue.

ClusterScripts.convert_to_grouped_jld2 — Method

convert_to_grouped_jld2(filename, results_data)

Converts the results format of an ungrouped JLD2 file to the grouped format. Warning: This method yields undefined simulation parameters which normally shouldn't occur in grouped JLD2 files.

ClusterScripts.create_results_file — Method

create_results_file(output_filename::String, glob_pattern::String, queue_file::String;trajectories_key="trajectories", truncate_times=true)

Compresses all results from a simulation queue back into a single file. Any missing outputs are reported as warnings and will be undefined in the final output.

This file contains the results of all jobs in the queue, as well as the input parameters for each job in the following format:

file["results"] contains an Array with the same dimensions as the input parameters.

Each element of file["results"] contains a tuple of the output from the simulation and the input parameters for that simulation.

Simulation output will always be a vector, even for single trajectories, to allow for consistent analysis functions that are independent of trajectory numbers.

Arguments output_filename::String: The name of the file to save the results to.

glob_pattern::String: A glob pattern to match all files to merge.

queue_file::String: The file containing the input parameters for the simulation queue.

trajectories_key::String: The key in the input parameters dictionary which describes batching behaviour. (typically "trajectories", since we want to farm out trajectories to workers)

truncate_times::Bool: If true, the time array in the output will be truncated to the final value only. Useful to save space when a large number of identical trajectories are run with short time steps.

ClusterScripts.merge_pmap_results — Method

merge_pmap_results(simulation_output;trajectories_key="trajectories", ensemble_key="ensemble_algorithm")
Merger function for NQCD simulation results. Takes a vector of (NQCD output, input_parameters) tuples and merges together all unique combinations of simulation parameters, adding the number of trajectories together as if the simulation was run as a larger ensemble.

TBW

ClusterScripts.pmap_queue — Method

pmap_queue(target_function::Function, input_list::Vector{Dict}; trajectories_key="trajectories", ensemble_key="ensemble_algorithm", tmp_dir="tmp/", checkpoint_frequency=0, sort_variable="")
Splits a list of input parameters into smaller operations for better multiprocessing, then runs the target function on the list of inputs and merges the results back together again.

Arguments similar to pmap: loadbalance_queue(f, args)

`trajectories_key`: Dictionary key in input parameter dictionary holding the number of trajectories for target_function to perform.

`ensemble_key`: Dictionary key in input parameters containing the EnsembleAlgorithm for NQCD to use.

`tmp_dir`: Location for checkpoint files.

`checkpoint_frequency`: Number of operations per worker to perform before saving to a tempfile. (Default is 1 per worker - the faster and more stable, the higher to go)

`sort_variable`: Offers the option to sort all jobs generated by a certain input parameter to group tasks of similar size together.

TBW

ClusterScripts.push_nqcd_outputs! — Method

push_nqcd_outputs!!(first_output, other_outputs...)

Like a push!() function, but it also puts `first_output` into a vector if it wasn't already and adds the number of trajectories together.

TBW

ClusterScripts.save! — Method

save!(loader::ResultsLazyLoader)

Updates the stored parameters and derived quantities inside a grouped JLD2 file. Run this function after modifying loader.parameters or loader.derived_quantities to save the changes.

ClusterScripts.serialise_queue! — Method

serialise_queue!(input_dict_tensor::Vector{<: Dict{<: Any}}; trajectories_key="trajectories", filename="simulation_parameters.jld2")

Performs batching on the Array of input parameters for multithreading/multiprocessing. By assigning the key "batchsize" in the input parameters, each simulation job will be split into as many batches as necessary to run the required number of trajectories. The default batch size is 1, i.e. trivial taskfarming.

Set "trajectories_key" in case jobs should be split by something different.

Set "filename" to save the resulting batch queue somewhere different than simulation_parameters.jld2.

ClusterScripts.update_results_file — Method

update_results_file(input_file::String, glob_pattern::String, queue_file::String, output_file::String; trajectories_key="trajectories", file_format::String="jld2")

Merges existing results from an ungrouped JLD2 file into a new ungrouped JLD2 file.

Arguments output_filename::String: The name of the file to save the results to.

glob_pattern::String: A glob pattern to match all files to merge.

queue_file::String: The file containing the input parameters for the simulation queue.

trajectories_key::String: The key in the input parameters dictionary which describes batching behaviour. (typically "trajectories", since we want to farm out trajectories to workers)